fix: validate if user exists before delete

+++ src/usecases/remove-user/usecase.ts
@@ -3,7 +3,6 @@ import { IRemoveUserGateway as G } from './gateway';
 import { IUsecase } from '@shared/protocols/usecase.protocol';
 import { Inject, Injectable } from '@nestjs/common';
 import { UserNotFoundError } from '@shared/errors';
-import { IUserEntity, User } from '@entities/user.entity';
 
 @Injectable()
 export class RemoveUserUsecase extends IUsecase<I, O, G> {
@@ -16,10 +15,14 @@ export class RemoveUserUsecase extends IUsecase<I, O, G> {
       throw new UserNotFoundError();
     }
 
-    const user = new User(
-      input.currentUser as Partial<IUserEntity>,
+    const user = await this.gateway.userRepository.findByPk(
       input.currentUser.id,
     );
+
+    if (!user?.id) {
+      throw new UserNotFoundError();
+    }
+
     const wasDeleted = await this.gateway.userRepository.remove(user);
 
     return { user, deleted: wasDeleted };
